Change Sort.svelte to use new sortable actions
